2.0.0 milestone changed to a flat function rather than an immutable class
making new instances of a DiceRoll that could only return the static result gained during construct led to awkward code that doesn't feel node-ish. The alternatives was to either make it rerollable, giving its class status some justification, or make it a flat function. In the end it didn't make sense to break immutability out of some arbitrary desire to make the thing OO, so it has been changed to a flat function for ease of use.